+++ nautilus_trader/backtest/matching_engine.pyx
@@ -1516,6 +1516,20 @@ cdef class OrderMatchingEngine:
 
         order.liquidity_side = liquidity_side
 
+        cdef:
+            Price fill_px
+            Quantity fill_qty
+            uint64_t total_size_raw = 0
+        if order.time_in_force == TimeInForce.FOK:
+            # Check FOK requirement
+            for fill in fills:
+                fill_px, fill_qty = fill
+                total_size_raw += fill_qty._mem.raw
+
+            if order.leaves_qty._mem.raw > total_size_raw:
+                self.cancel_order(order)
+                return  # Cannot fill full size - so kill/cancel
+
         if not fills:
             self._log.error(
                 "Cannot fill order: no fills from book when fills were expected (check sizes in data).",
@@ -1534,8 +1548,6 @@ cdef class OrderMatchingEngine:
             )
 
         cdef:
-            Price fill_px
-            Quantity fill_qty
             bint initial_market_to_limit_fill = False
             Price last_fill_px = None
         for fill_px, fill_qty in fills:
@@ -1548,14 +1560,6 @@ cdef class OrderMatchingEngine:
                         trigger_price=None,
                     )
                     initial_market_to_limit_fill = True
-                if order.time_in_force == TimeInForce.FOK and fill_qty._mem.raw < order.quantity._mem.raw:
-                    # FOK order cannot fill the entire quantity - cancel
-                    self.cancel_order(order)
-                    return
-            elif order.time_in_force == TimeInForce.IOC:
-                # IOC order has already filled at one price - cancel remaining
-                self.cancel_order(order)
-                return
 
             if self.book_type == BookType.L1_MBP and self._fill_model.is_slipped():
                 if order.side == OrderSide.BUY:
@@ -1598,6 +1602,11 @@ cdef class OrderMatchingEngine:
 
             last_fill_px = fill_px
 
+        if order.time_in_force == TimeInForce.IOC and order.is_open_c():
+            # IOC order has filled all available size
+            self.cancel_order(order)
+            return
+
         if (
             order.is_open_c()
             and self.book_type == BookType.L1_MBP
@@ -1607,11 +1616,6 @@ cdef class OrderMatchingEngine:
             or order.order_type == OrderType.STOP_MARKET
         )
         ):
-            if order.time_in_force == TimeInForce.IOC:
-                # IOC order has already filled at one price - cancel remaining
-                self.cancel_order(order)
-                return
-
             # Exhausted simulated book volume (continue aggressive filling into next level)
             # This is a very basic implementation of slipping by a single tick, in the future
             # we will implement more detailed fill modeling.
@@ -1632,6 +1636,7 @@ cdef class OrderMatchingEngine:
                 venue_position_id=venue_position_id,
                 position=position,
             )
+
 
     cpdef void fill_order(
         self,